mainmenue 'goto' entfernt/Kommentare eingefügt/Klasse separiert
